Use upstream modification date as internal_version
When using the latest pointer, no version is configured in images.yml.
Hence we retrieve the last modification date of the upstream image file
and set it as the internal_version. When no upstream date can be
retrieved, temporarily set the internal_version to 'latest' and later use
the creation date from Openstack when renaming the image.

Signed-off-by: Simon Hensel <simon.hensel@protonmail.com>