Refactor & implement unittest for manage.py (#236)
Refactor manage.py into smaller methods to increase readability and maintainability.
All functions now have proper docstrings that indicate argument and return types.

Implement a unittest for manage.py that tests all functions in
isolation against a mocked glance API.
The tests can be run with `tox -e test`.

The --use-raw-image flag is removed, as it leads to errors and inconsistencies
when the image conversion is not enabled in Glance or an image format
gets uploaded that cannot be converted (e.g. ISO). Glance itself sets the image
format to RAW after the conversion, regardless of the previously configured format.

Closes #227 and #229

Signed-off-by: Simon Hensel <simon.hensel@inovex.de>